Closes #692 added mod asset interception support
