+++ src/SMAPI/Framework/ModHelpers/ContentHelper.cs
@@ -79,6 +79,30 @@ namespace StardewModdingAPI.Framework.ModHelpers
             this.Monitor = monitor;
         }
 
+        /// <summary>Load content from the mod folder (if not already cached), and return it. When loading a <c>.png</c> file, this must be called outside the game's draw loop.</summary>
+        /// <typeparam name="T">The expected data type. The main supported types are <see cref="Map"/>, <see cref="Texture2D"/>, and dictionaries; other types may be supported by the game's content pipeline.</typeparam>
+        /// <param name="key">The asset local path to a content file relative to the mod folder.</param>
+        /// <param name="allowModEdits">Allow asset to be intercepted by another mods.</param>
+        /// <exception cref="ArgumentException">The <paramref name="key"/> is empty or contains invalid characters.</exception>
+        /// <exception cref="ContentLoadException">The content asset couldn't be loaded (e.g. because it doesn't exist).</exception>
+        public T Load<T>(string key, bool allowModEdits)
+        {
+            if (!allowModEdits)
+            {
+                return this.Load<T>(key, ContentSource.ModFolder);
+            }
+
+            try
+            {
+                this.AssertAndNormalizeAssetName(key);
+                return this.ModContentManager.LoadWithEdits<T>(key, this.ModID, this.CurrentLocaleConstant, useCache: false);
+            }
+            catch (Exception ex)
+            {
+                throw new SContentLoadException($"{this.ModName} failed loading content asset '{key}' from {ContentSource.ModFolder}.", ex);
+            }
+        }
+
         /// <summary>Load content from the game folder or mod folder (if not already cached), and return it. When loading a <c>.png</c> file, this must be called outside the game's draw loop.</summary>
         /// <typeparam name="T">The expected data type. The main supported types are <see cref="Map"/>, <see cref="Texture2D"/>, and dictionaries; other types may be supported by the game's content pipeline.</typeparam>
         /// <param name="key">The asset key to fetch (if the <paramref name="source"/> is <see cref="ContentSource.GameContent"/>), or the local path to a content file relative to the mod folder.</param>
